+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Notable `minecraft-status` changes, tracked in the [keep a changelog](https://ke
 
 ## [Unreleased]
 
+### Fixed
+
+* "Widget Setup" now links to the iPad widget help article on iPads
+
 ## [v0.1.0-5] - 2021-04-14
 
 This release adds support for pinging Bedrock Minecraft servers! The widgets default to an "Auto" mode that can ping any Minecraft server type successfully, and you can choose a specific type in the widget's settings as well if necessary.